+++ sdk/core/core-client/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added a new function `createCAECallbacks`, that generates callbacks to use on the `bearerTokenAuthenticationPolicy` to support [Continuous Access Evaluation (CAE) challenges](https://docs.microsoft.com/azure/active-directory/conditional-access/concept-continuous-access-evaluation).
+- Added a new function `authorizeRequestOnClaimChallenge`, that can be used with the `@azure/core-rest-pipeline`'s `bearerTokenAuthenticationPolicy` to support [Continuous Access Evaluation (CAE) challenges](https://docs.microsoft.com/azure/active-directory/conditional-access/concept-continuous-access-evaluation).
 
 ### Breaking Changes